+++ src/aqua/graphics/timeseries.py
@@ -5,21 +5,28 @@ both with monthly and annual aggregation options
 import xarray as xr
 import matplotlib.pyplot as plt
 from aqua.logger import log_configure
-from .util_timeseries import plot_monthly_data, plot_annual_data, plot_ref_monthly_data, plot_ref_annual_data
+from .util_timeseries import plot_timeseries_data, plot_timeseries_ref_data, plot_timeseries_ensemble
 from .styles import ConfigStyle
 
 
-def plot_timeseries(monthly_data=None,
-                    annual_data=None,
-                    ref_monthly_data=None,
-                    ref_annual_data=None,
-                    std_monthly_data=None,
-                    std_annual_data=None,
+def plot_timeseries(monthly_data: list[xr.DataArray] | xr.DataArray = None,
+                    annual_data: list[xr.DataArray] | xr.DataArray = None,
+                    ref_monthly_data: xr.DataArray = None,
+                    ref_annual_data: xr.DataArray = None,
+                    std_monthly_data: xr.DataArray = None,
+                    std_annual_data: xr.DataArray = None,
+                    ens_monthly_data: xr.DataArray = None,
+                    ens_annual_data: xr.DataArray = None,
+                    std_ens_monthly_data: xr.DataArray = None,
+                    std_ens_annual_data: xr.DataArray = None,
                     data_labels: list = None,
                     ref_label: str = None,
+                    ens_label: str = None,
                     style: str = None,
                     fig: plt.Figure = None,
                     ax: plt.Axes = None,
+                    figsize: tuple = (10, 5),
+                    title: str = None,
                     loglevel: str = 'WARNING',
                     **kwargs):
     """
@@ -34,11 +41,17 @@ def plot_timeseries(monthly_data=None,
         ref_annual_data (xr.DataArray): reference annual data to plot
         std_monthly_data (xr.DataArray): standard deviation of the reference monthly data
         std_annual_data (xr.DataArray): standard deviation of the reference annual data
+        ens_monthly_data (xr.DataArray): ensemble monthly data to plot
+        ens_annual_data (xr.DataArray): ensemble annual data to plot
+        std_ens_monthly_data (xr.DataArray): standard deviation of the ensemble monthly data
+        std_ens_annual_data (xr.DataArray): standard deviation of the ensemble annual data
         data_labels (list of str): labels for the data
         ref_label (str): label for the reference data
         style (str): style to use for the plot. By default the schema specified in the configuration file is used.
         fig (plt.Figure): figure object to plot on
         ax (plt.Axes): axis object to plot on
+        figsize (tuple): size of the figure
+        title (str): title of the plot
         loglevel (str): logging level
 
     Keyword Arguments:
@@ -48,30 +61,51 @@ def plot_timeseries(monthly_data=None,
     Returns:
         fig, ax (tuple): tuple containing the figure and axis objects
     """
-    logger = log_configure(loglevel, 'PlotTimeseries')
+    logger = log_configure(loglevel, 'plot_timeseries')
     ConfigStyle(style=style, loglevel=loglevel)
+    realization = False
 
     if fig is None and ax is None:
-        fig_size = kwargs.get('figsize', (10, 5))
-        fig, ax = plt.subplots(1, 1, figsize=fig_size)
+        fig, ax = plt.subplots(1, 1, figsize=figsize)
 
+    if (monthly_data is not None and ens_monthly_data is not None) or (annual_data is not None and ens_annual_data is not None):
+        logger.info("monthly_data and annual_data will be considered as realizations of an ensemble")
+        realization = True
 
     if monthly_data is not None:
-        plot_monthly_data(ax, monthly_data, data_labels, logger, lw=3)
+        plot_timeseries_data(ax=ax, data=monthly_data, kind='monthly',
+                             data_labels=data_labels, realization=realization,
+                             lw=2.5 if not realization else 0.8)
 
     if annual_data is not None:
-        plot_annual_data(ax, annual_data, data_labels, logger, lw=3)
+        plot_timeseries_data(ax=ax, data=annual_data, kind='annual',
+                             data_labels=data_labels, realization=realization,
+                             lw=2.5 if not realization else 0.8)
 
     if ref_monthly_data is not None:
-        plot_ref_monthly_data(ax, ref_monthly_data, std_monthly_data, ref_label, logger, lw=0.8)
+        plot_timeseries_ref_data(ax=ax, ref_data=ref_monthly_data,
+                                 std_data=std_monthly_data,
+                                 ref_label=ref_label, lw=0.8, kind='monthly')
 
     if ref_annual_data is not None:
-        plot_ref_annual_data(ax, ref_annual_data, std_annual_data, ref_label, logger, lw=0.8)
+        plot_timeseries_ref_data(ax=ax, ref_data=ref_annual_data,
+                                 std_data=std_annual_data,
+                                 ref_label=ref_label, lw=0.8, kind='annual')
     
-    ax.legend(fontsize='small')
+    if ens_monthly_data is not None:
+        plot_timeseries_ensemble(ax=ax, data=ens_monthly_data,
+                                 std_data=std_ens_monthly_data,
+                                 data_label=ens_label, lw=0.8, kind='monthly')
+    
+    if ens_annual_data is not None:
+        plot_timeseries_ensemble(ax=ax, data=ens_annual_data,
+                                 std_data=std_ens_annual_data,
+                                 data_label=ens_label, lw=0.8, kind='annual')
+    
+    if data_labels is not None or ref_label is not None or ens_label is not None:
+        ax.legend(fontsize='small')
     ax.grid(True, axis="y", linestyle='-', color='silver', alpha=0.8)
 
-    title = kwargs.get('title', None)
     if title:
         ax.set_title(title, fontsize=13, fontweight='bold')